requirement.txt
alembic==1.4.2
backcall==0.2.0
blinker==1.4
boto3==1.14.56
boto3-stubs==1.14.56.0
botocore==1.17.56
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
docutils==0.15.2
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
mysqlclient==2.0.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
razorpay==1.2.0
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-ImageAttach==1.1.0
traitlets==4.3.3
twilio==6.45.4
typing-extensions==3.7.4.3
urllib3==1.25.10
Wand==0.6.2
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
